refactor(tasklist): remove graphQL client
Implement Tasklist REST API only
