+++ .vscode/settings.json
@@ -1,6 +1,7 @@
 {
     "conventionalCommits.scopes": [
         "repo",
-        "oauth"
+        "oauth",
+        "tasklist"
     ]
 }
