+++ packages/tasklist/src/__test__/test.spec.ts
@@ -1,5 +1,4 @@
-import { TaskState } from '../lib/Types'
-import { TasklistApiClient } from './../index'
+import { TasklistApiClient } from '../index'
 import { join } from 'path'
 import { CreateProcessInstanceResponse, DeployProcessResponse, ZBClient } from '@jwulf/zeebe'
 import 'dotenv/config'
@@ -10,6 +9,7 @@ let p: CreateProcessInstanceResponse
 let def: DeployProcessResponse
 
 const delay = (ms: number) => new Promise((resolve) => setTimeout(() => resolve(null), ms))
+
 describe('TasklistApiClient', () => {
     const zbc = new ZBClient({
         loglevel: 'NONE',
@@ -40,147 +40,99 @@ describe('TasklistApiClient', () => {
 
     afterAll(() => zbc.close())
 
-    it('can request all tasks', async () => {
-        const tasklist = new TasklistApiClient()
-        const { tasks } = await tasklist.getAllTasks(['id', 'processDefinitionId', 'taskDefinitionId'])
-        expect(tasks.length).toBeGreaterThan(0)
-    })
-
-    it('can request a task with parameters', async () => {
-        const tasklist = new TasklistApiClient()
-        const { tasks } = await tasklist.getTasks({
-            state: TaskState.CREATED,
+    describe('Read operations', () => {
+        it('can request all tasks', async () => {
+            const tasklist = new TasklistApiClient()
+            const tasks = await tasklist.getAllTasks()
+            expect(tasks.length).toBeGreaterThan(0)
         })
-        expect(tasks.length).toBeGreaterThan(0)
-    })
 
-    it('can request tasks with fields', async () => {
-        const tasklist = new TasklistApiClient()
-        const { tasks } = await tasklist.getTasks(
-            {
-                state: TaskState.CREATED,
-            },
-            ['id', 'processName']
-        )
-        expect(tasks.length).toBeGreaterThan(0)
-        expect(tasks[0].name).not.toBeDefined()
-        expect(tasks[0].processName).toBeTruthy()
-    })
-
-    it('can request tasks with all fields', async () => {
-        const tasklist = new TasklistApiClient()
-        const { tasks } = await tasklist.getTasks({
-            state: TaskState.CREATED,
+        it('can request a task with parameters', async () => {
+            const tasklist = new TasklistApiClient()
+            const tasks = await tasklist.getTasks({
+                state: 'CREATED',
+            })
+            expect(tasks.length).toBeGreaterThan(0)
         })
-        expect(tasks.length).toBeGreaterThan(0)
-        expect(tasks[0].processName).toBeTruthy()
-    })
 
-    it('can request a specific task', async () => {
-        const tasklist = new TasklistApiClient()
-        const { tasks } = await tasklist.getTasks({
-            state: TaskState.CREATED,
+        it('gets all fields for a task', async () => {
+            const tasklist = new TasklistApiClient()
+            const tasks = await tasklist.getTasks({
+                state: 'CREATED',
+            })
+            expect(tasks.length).toBeGreaterThan(0)
+            expect(tasks[0].name).toBeTruthy()
+            expect(tasks[0].processName).toBeTruthy()
         })
-        const id = tasks[0].id
-        const { task } = await tasklist.getTask(id)
-        expect(task.id).toBe(id)
+
+        it('can request a specific task', async () => {
+            const tasklist = new TasklistApiClient()
+            const tasks = await tasklist.getTasks({
+                state: 'CREATED',
+            })
+            const id = tasks[0].id
+            const task = await tasklist.getTask(id)
+            expect(task.id).toBe(id)
+        })
+
+        it('can retrieve an embedded form', async () => {
+            const tasklist = new TasklistApiClient()
+            const res = await tasklist.getForm('userTaskForm_3r97fja', def.processes[0].processDefinitionKey)
+            expect(res.id).toBe('userTaskForm_3r97fja')
+        })
     })
 
-    it('can retrieve an embedded form', async () => {
-        const tasklist = new TasklistApiClient()
-        const res = await tasklist.getForm('userTaskForm_3r97fja', def.processes[0].processDefinitionKey)
-        expect(res.form.id).toBe('userTaskForm_3r97fja')
-    })
+    describe('Write operations', () => {
+        it('can claim a task', async () => {
+            const tasklist = new TasklistApiClient()
+            const tasks = await tasklist.getTasks({ state: 'CREATED' })
+            const taskid = tasks[0].id
+            expect(tasks.length).toBeGreaterThan(0)
+            const claimTask = await tasklist.assignTask({ taskId: taskid, assignee: 'jwulf' })
+            expect(claimTask.id).toBe(taskid)
+        })
 
-    xit('can retrieve the current logged in user', async () => {
-        const tasklist = new TasklistApiClient()
-        const res = await tasklist.getCurrentUser()
-        expect(res).toBeTruthy()
-    })
+        it('will not allow a task to be claimed twice', async () => {
+            const tasklist = new TasklistApiClient()
+            const tasks = await tasklist.getTasks({ state: 'CREATED' })
+            const task = await tasklist.assignTask({ taskId: tasks[0].id, assignee: 'jwulf' })
+            expect(task).toBeTruthy()
+            let threw = false
+            try {
+                await tasklist.assignTask({ taskId: tasks[0].id, assignee: 'jwulf0', allowOverrideAssignment: false })
+            } catch (e) {
+                threw = true
+            }
+            expect(threw).toBe(true)
+        })
 
-    it('can claim a task', async () => {
-        const tasklist = new TasklistApiClient()
-        const { tasks } = await tasklist.getTasks({ state: TaskState.CREATED })
-        const taskid = tasks[0].id
-        expect(tasks.length).toBeGreaterThan(0)
-        const { claimTask } = await tasklist.claimTask(taskid, 'jwulf')
-        expect(claimTask.id).toBe(taskid)
-    })
+        it('can unclaim task', async () => {
+            const tasklist = new TasklistApiClient()
+            const tasks = await tasklist.getTasks({ state: 'CREATED' })
+            const taskId = tasks[0].id
+            const task = await tasklist.assignTask({ taskId: taskId, assignee: 'jwulf', allowOverrideAssignment: false })
+            expect(task).toBeTruthy()
+            let threw = false
+            try {
+                await tasklist.assignTask({ taskId: taskId, assignee: 'jwulf0', allowOverrideAssignment: false })
+            } catch {
+                threw = true
+            }
+            expect(threw).toBe(true)
+            await tasklist.unassignTask(taskId)
+            const claimTask = await tasklist.assignTask({ taskId, assignee: 'jwulf0', allowOverrideAssignment: false })
+            expect(claimTask.id).toEqual(taskId)
+        })
 
-    it('will not allow a task to be claimed twice', async () => {
-        const tasklist = new TasklistApiClient()
-        const tasks = await tasklist.getTasks({ state: TaskState.CREATED })
-        const task = await tasklist.claimTask(tasks.tasks[0].id, 'jwulf')
-        expect(task).toBeTruthy()
-        let threw = false
-        try {
-            await tasklist.claimTask(tasks.tasks[0].id, 'jwulf0', false)
-        } catch (e) {
-            threw = true
-        }
-        expect(threw).toBe(true)
-    })
-
-    it('can unclaim task', async () => {
-        const tasklist = new TasklistApiClient()
-        const tasks = await tasklist.getTasks({ state: TaskState.CREATED })
-        const taskid = tasks.tasks[0].id
-        const task = await tasklist.claimTask(taskid, 'jwulf')
-        expect(task).toBeTruthy()
-        let threw = false
-        try {
-            await tasklist.claimTask(taskid, 'jwulf0', false)
-        } catch {
-            threw = true
-        }
-        expect(threw).toBe(true)
-        await tasklist.unclaimTask(taskid)
-        const res = await tasklist.claimTask(taskid, 'jwulf0', false)
-        expect(res.claimTask.id).toEqual(taskid)
-    })
-
-    it('can delete a process instance', async () => {
-        const tasklist = new TasklistApiClient()
-        const { tasks } = await tasklist.getTasks({ state: TaskState.COMPLETED }, ['processInstanceId'])
-        expect(tasks.length).toBeGreaterThan(0)
-        const id = tasks[0].processInstanceId
-        const { deleteProcessInstance } = await tasklist.deleteProcessInstance(id)
-        expect(deleteProcessInstance).toBe(true)
-    })
-
-    it('can complete a Task', async () => {
-        const tasklist = new TasklistApiClient()
-        const { tasks } = await tasklist.getTasks({ state: TaskState.CREATED })
-        const taskid = tasks[0].id
-        expect(tasks.length).toBeGreaterThan(0)
-        const { completeTask } = await tasklist.completeTask(taskid, { outcome: 'approved', fruits: ['apple', 'orange'] })
-        expect(completeTask.id).toBe(taskid)
-        p = null as any
-    })
-
-    it('can complete a task with variables', (done) => {
-        const tasklist = new TasklistApiClient()
-        zbc.createProcessInstanceWithResult({
-            bpmnProcessId: 'TasklistTestProcess',
-            variables: {
-                name: 'Joe Bloggs',
-                age: 42,
-                interests: ['golf', 'frisbee'],
-            },
-        }).then((res) => {
-            console.log(res)
+        it('can complete a Task', async () => {
+            const tasklist = new TasklistApiClient()
+            const tasks = await tasklist.getTasks({ state: 'CREATED' })
+            const taskid = tasks[0].id
+            expect(tasks.length).toBeGreaterThan(0)
+            const completeTask = await tasklist.completeTask(taskid, { outcome: 'approved', fruits: ['apple', 'orange'] })
+            expect(completeTask.id).toBe(taskid)
+            expect(completeTask.taskState).toEqual('COMPLETED')
             p = null as any
-            done()
         })
-        delay(10000)
-            .then(() => tasklist.getTasks({ state: TaskState.CREATED }))
-            .then(({ tasks }) =>
-                tasks.forEach((task) =>
-                    tasklist.completeTask(task.id, {
-                        outcome: 'approved',
-                        fruits: ['apple', 'orange'],
-                    })
-                )
-            )
     })
 })
