+++ packages/tasklist/src/index.ts
@@ -1,4 +1,3 @@
 export { TasklistApiClient } from './lib/TasklistApiClient'
-export { TasklistRESTClient } from './lib/TasklistRESTClient'
 
 export * from './lib/Types'
