+++ packages/tasklist/src/lib/TasklistApiClient.ts
@@ -1,30 +1,15 @@
-import { getTasklistToken, getTasklistCredentials } from '@jwulf/oauth'
-import gotQl from 'gotql'
-import { Form, GraphQLTasksQuery, Task, TaskFields, TaskQuery, TaskWithVariables, User, Variable } from './Types'
-import { getResponseDataOrThrow, decodeTaskVariablesFromGraphQL, encodeTaskVariablesForAPIRequest, JSONDoc } from './utils'
+import { OAuthProviderImpl, getTasklistToken, getTasklistCredentials } from '@jwulf/oauth'
+import got from 'got'
+import { Form, Task, TaskQuery, Variable } from './Types'
+import { encodeTaskVariablesForAPIRequest, JSONDoc } from './utils'
+import { debug } from 'debug'
+const trace = debug('tasklist:rest')
 
 const pkg = require('../../package.json')
-
-const defaultFields: TaskFields = [
-    'assignee',
-    'candidateGroups',
-    'completionTime',
-    'creationTime',
-    'formKey',
-    'id',
-    'isFirst',
-    'name',
-    'processDefinitionId',
-    'processInstanceId',
-    'processName',
-    'sortValues',
-    'taskDefinitionId',
-    'taskState',
-    { variables: { fields: ['name', 'value'] } },
-] as any
+const TASKLIST_API_VERSION = 'v1'
 
 /**
- * @description The high-level client for the Tasklist GraphQL API
+ * @description The high-level client for the Tasklist REST API
  * @example
  * ```
  *
@@ -32,7 +17,8 @@ const defaultFields: TaskFields = [
  */
 export class TasklistApiClient {
     private userAgentString: string
-    graphqlUrl: string
+    private oauthProvider: OAuthProviderImpl | undefined
+    private rest: typeof got
 
     /**
      * @example
@@ -42,16 +28,29 @@ export class TasklistApiClient {
      * @description
      *
      */
-    constructor() {
-        this.userAgentString = `tasklist-graphql-client-nodejs/${pkg.version}`
+    constructor(
+        options: {
+            oauthProvider?: OAuthProviderImpl
+            baseUrl?: string
+        } = {}
+    ) {
+        this.oauthProvider = options.oauthProvider
+        this.userAgentString = `tasklist-rest-client-nodejs/${pkg.version}`
         const creds = getTasklistCredentials()
-        this.graphqlUrl = `${creds.CAMUNDA_TASKLIST_BASE_URL}/graphql`
+        const baseUrl = options.baseUrl ?? `${creds.CAMUNDA_TASKLIST_BASE_URL}`
+        const prefixUrl = `${baseUrl}/${TASKLIST_API_VERSION}`
+        this.rest = got.extend({
+            prefixUrl,
+        })
+        trace(`prefixUrl: ${prefixUrl}`)
     }
 
     private async getHeaders() {
+        const token = this.oauthProvider ? await this.oauthProvider.getToken('TASKLIST') : await getTasklistToken(this.userAgentString)
+        trace(`Got a token: ${token}`)
         return {
             'content-type': 'application/json',
-            authorization: `Bearer ${await getTasklistToken(this.userAgentString)}`,
+            authorization: `Bearer ${token}`,
             'user-agent': this.userAgentString,
             accept: '*/*',
         }
@@ -65,199 +64,142 @@ export class TasklistApiClient {
      * async function getTasks() {
      *   const res = await tasklist.getTasks({
      *     state: TaskState.CREATED
-     *   }, ['id', 'name', 'processName'])
-     *   console.log(res ? 'Nothing' : JSON.stringify(res.tasks, null, 2))
+     *   })
+     *   console.log(res ? 'Nothing' : JSON.stringify(res, null, 2))
      *   return res
      * }
      * ```
      * @param query
-     * @param fields - a list of fields to return in the query results
-     *
      */
-    public async getTasks<T = { [key: string]: any }>(
-        query: Partial<TaskQuery>,
-        fields: TaskFields = defaultFields
-    ): Promise<{ tasks: TaskWithVariables<T>[] }> {
+    public async getTasks(query: Partial<TaskQuery>): Promise<Task[]> {
         const headers = await this.getHeaders()
-        const q: GraphQLTasksQuery = {
-            operation: {
-                name: 'tasks',
-                args: {
-                    query,
-                },
-                fields,
-            },
-        }
-        return gotQl
-            .query(this.graphqlUrl, q, { headers })
-            .then((res) => ({ tasks: getResponseDataOrThrow<{ tasks: Task[] }>(res).tasks.map(decodeTaskVariablesFromGraphQL<T>) }))
+        const url = 'tasks/search'
+        trace(`Requesting ${url}`)
+        return this.rest
+            .post(url, {
+                json: query,
+                headers,
+            })
+            .json()
     }
 
-    public async getAllTasks<T = { [key: string]: any }>(fields: TaskFields = defaultFields): Promise<{ tasks: TaskWithVariables<T>[] }> {
-        return this.getTasks<T>({}, fields)
+    public async getAllTasks(): Promise<Task[]> {
+        return this.getTasks({})
     }
 
     /**
-     * @description https://docs.camunda.io/docs/apis-clients/tasklist-api/queries/task/
+     * @description Return a task by id, or throw if not found.
+     * @throws Will throw if no task of the given id exists
      * @param id
-     * @param fields
      * @returns
      */
-    public async getTask<T = { [key: string]: any }>(id: string, fields = defaultFields): Promise<{ task: TaskWithVariables<T> }> {
+    public async getTask(taskId: string): Promise<Task> {
         const headers = await this.getHeaders()
-        const query = {
-            operation: {
-                name: 'task',
-                args: {
-                    id,
-                },
-                fields,
-            },
-        }
-        return gotQl
-            .query(this.graphqlUrl, query, { headers })
-            .then((res) => ({ task: decodeTaskVariablesFromGraphQL<T>(getResponseDataOrThrow<{ task: Task }>(res).task) }))
+        return this.rest
+            .get(`tasks/${taskId}`, {
+                headers,
+            })
+            .json()
     }
 
     /**
-     * @description https://docs.camunda.io/docs/apis-clients/tasklist-api/queries/form/
-     * @param id
-     * @param processDefinitionId
+     * @description Get the form details by form id and processDefinitionKey.
+     * @param formId
+     * @param processDefinitionKey
      */
-    public async getForm(id: string, processDefinitionId: string): Promise<{ form: Form }> {
+    public async getForm(formId: string, processDefinitionKey: string): Promise<Form> {
         const headers = await this.getHeaders()
-        const query = {
-            operation: {
-                name: 'form',
-                args: {
-                    id,
-                    processDefinitionId,
+        return this.rest
+            .get(`forms/${formId}`, {
+                searchParams: {
+                    processDefinitionKey,
                 },
-                fields: ['id', 'processDefinitionId', 'schema'],
-            },
-        }
-        return gotQl.query(this.graphqlUrl, query, { headers }).then((res) => getResponseDataOrThrow(res))
+                headers,
+            })
+            .json()
     }
 
     /**
-     * @description https://docs.camunda.io/docs/apis-clients/tasklist-api/queries/current-user/
-     */
-    public async getCurrentUser(): Promise<User> {
-        const headers = await this.getHeaders()
-        const query = {
-            operation: {
-                name: 'user',
-                fields: ['userId', 'displayName', 'permissions', 'roles', 'salesPlanType'],
-            },
-        }
-        return gotQl.query(this.graphqlUrl, query, { headers }).then((res) => getResponseDataOrThrow(res))
-    }
-
-    /**
-     * @description https://docs.camunda.io/docs/apis-clients/tasklist-api/queries/variables/
+     * @description Returns a list of variables
      * @param taskId
      * @param variableNames
+     * @throws Throws 404 if no task of taskId is found
      */
-    public async getVariables(taskId: string, variableNames: string[]) {
-        throw new Error(`Not implemented yet ${taskId} ${variableNames}`)
+    public async getVariables(taskId: string, variableNames?: string[]): Promise<Variable[]> {
+        const headers = await this.getHeaders()
+        return this.rest
+            .post(`tasks/${taskId}/variables/search`, {
+                body: JSON.stringify(variableNames || []),
+                headers,
+            })
+            .json()
     }
 
     /**
      * @description https://docs.camunda.io/docs/apis-clients/tasklist-api/queries/variable/
      * @param id
+     * @throws Throws 404 if no variable of the id is found
      */
-    public async getVariable(id: string): Promise<Variable> {
-        throw new Error(`Not implemented yet ${id}`)
+    public async getVariable(variableId: string): Promise<Variable> {
+        const headers = await this.getHeaders()
+        return this.rest
+            .get(`variables/${variableId}`, {
+                headers,
+            })
+            .json()
     }
 
     /**
-     * @description https://docs.camunda.io/docs/apis-clients/tasklist-api/mutations/claim-task/
+     * @description Assign a task with taskId to assignee or the active user.
      * @param taskId
-     * @param assignee
+     * @param assignee if not provided, assigns to the user whose JWT is used
      * @param allowOverrideAssignment
+     * @throws 400 - task not active, or already assigned. 403 - no permission to reassign task. 404 - no task for taskId.
      */
-    public async claimTask(taskId: string, assignee: string, allowOverrideAssignment: boolean = true): Promise<{ claimTask: Task }> {
+    public async assignTask({
+        taskId,
+        allowOverrideAssignment = false,
+        assignee,
+    }: {
+        taskId: string
+        assignee?: string
+        allowOverrideAssignment?: boolean
+    }): Promise<Task> {
         const headers = await this.getHeaders()
-        const query = {
-            operation: {
-                name: 'claimTask',
-                args: {
-                    taskId,
+        return this.rest
+            .patch(`tasks/${taskId}/assign`, {
+                body: JSON.stringify({
                     assignee,
                     allowOverrideAssignment,
-                },
-                fields: defaultFields,
-            },
-        }
-        return gotQl
-            .mutation(
-                this.graphqlUrl,
-                query as any, // the typing seems to have an error
-                { headers }
-            )
-            .then((res) => getResponseDataOrThrow(res))
+                }),
+                headers,
+            })
+            .json()
     }
 
     /**
-     * @description https://docs.camunda.io/docs/apis-clients/tasklist-api/mutations/complete-task/
+     * @description Complete a task with taskId and optional variables
      * @param taskId
      * @param variables
      */
-    public async completeTask(taskId: string, variables: JSONDoc): Promise<{ completeTask: Task }> {
+    public async completeTask(taskId: string, variables?: JSONDoc): Promise<Task> {
         const headers = await this.getHeaders()
-        const query = {
-            operation: {
-                name: 'completeTask',
-                args: {
-                    taskId,
-                    variables: '$completionVariables',
-                },
-                fields: defaultFields,
-            },
-            variables: {
-                completionVariables: {
-                    type: '[VariableInput!]!',
-                    value: encodeTaskVariablesForAPIRequest(variables),
-                },
-            },
-        }
-        return gotQl.mutation(this.graphqlUrl, query as any, { headers }).then((res) => getResponseDataOrThrow(res))
+        return this.rest
+            .patch(`tasks/${taskId}/complete`, {
+                headers,
+                body: JSON.stringify({
+                    variables: encodeTaskVariablesForAPIRequest(variables || {}),
+                }),
+            })
+            .json()
     }
 
     /**
-     * @description Delete process instance data from the Tasklist ES by id. Returns true if the process instance is found and canceled, false if the process instance could not be found.
-     * [Documentation](https://docs.camunda.io/docs/apis-clients/tasklist-api/mutations/delete-process-instance/).
-     * @param processInstanceId
-     */
-    public async deleteProcessInstance(processInstanceId: string): Promise<{ deleteProcessInstance: boolean }> {
-        const headers = await this.getHeaders()
-        const query = {
-            operation: {
-                name: 'deleteProcessInstance',
-                args: {
-                    processInstanceId,
-                },
-                fields: [],
-            },
-        }
-        return gotQl.mutation(this.graphqlUrl, query, { headers }).then((res) => getResponseDataOrThrow(res))
-    }
-
-    /**
-     * @description https://docs.camunda.io/docs/apis-clients/tasklist-api/mutations/unclaim-task/
+     * @description Unassign a task with taskI
      * @param taskId
      */
-    public async unclaimTask(taskId: string): Promise<Task> {
+    public async unassignTask(taskId: string): Promise<Task> {
         const headers = await this.getHeaders()
-        const query = {
-            operation: {
-                name: 'unclaimTask',
-                args: {
-                    taskId,
-                },
-                fields: defaultFields,
-            },
-        }
-        return gotQl.mutation(this.graphqlUrl, query, { headers }).then((res) => getResponseDataOrThrow(res))
+        return this.rest.patch(`tasks/${taskId}/unassign`, { headers }).json()
     }
 }
